requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
aioice==0.7.5
aiortc==1.2.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
av==8.0.3
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
crc32c==2.2.post0
cryptography==3.4.7
dataclasses==0.8
decorator==5.0.9
defusedxml==0.7.1
dnspython==2.1.0
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.18
idna==2.10
importlib-metadata==4.5.0
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.11.0
notebook==6.4.0
numpy==1.19.5
opencv-python==4.5.2.54
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyarrow==4.0.1
pycparser==2.20
pydeck==0.6.2
pyee==8.1.0
Pygments==2.9.0
pylibsrtp==0.6.8
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
Send2Trash==1.7.1
six==1.16.0
smmap==4.0.0
streamlit==0.83.0
streamlit-webrtc==0.22.3
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
torch==1.8.1+cpu
traitlets==4.3.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.1
opencv-python-headless==4.2.0.32